refactor resource module
